Implemented integer-RPN collapsing function; simplifies RPN expressions.
This new function is a friend of the RPNNotation class, takes in an
RPNNotation instance that contains a converted RPN notation expression,
and simplifies the sections where known integer constants exist,
skipping past variables. The result is that an expression containing
solely integer constants should be collapsed down to one integer. If
there are variables in the expression, they are simply ignored and the
expression is collapsed as much as possible. Note that this function is
completely untested, and the next thing I should do is write a testing
program for this function. Only an integer expression collapsing
function was written, meaning a floating point expression collapsing
function still needs to be written.