Wrote a testing program for the collapseRPNI function.
Demonstrates proper behavior for both pure-constant arithmetic
expressions and expressions with variables in them. Note that this
function assumes correct RPN format, and cannot handle incorrect
formats. That said, these functions are linearly linked in such a way
that a higher level function *should* never be fed incorrect input from a
lower level function if it assumes the format is correct.