Modified commitment Makefile setup.
Deleted the Commit.sh and Pull.sh scripts in favor of specialized
CommitX.sh, PullX.sh, and PanicX.sh scripts, where X is some suffix that
represents the branch we would like to operate on. For example,
CommitM.sh commits and pushes the master branch to GitHub, PullM.sh
pulls the master branch from GitHub, and PanicM.sh reverts to before the
last pull from the master branch on GitHub.