Modified pulling scripts and added remote tracking setup script.
Each of the Pull scripts does literally exactly the same thing and I
will probably revert back to one pull script in a moment. However, the
SetupRemoteTracking.sh script is important. This script should be
updated every time a new branch is made that needs to be linked with a
new remote branch, and once this script is run, the current git system
will be able to correctly pull whatever remote branch is related to
whatever the local branch is. The --set-upstream flag is the flag that
asocciates local and remote branches.