Updated the lexing algorithm to provide more useful output.
A modified lexing algorithm, tokenizeFileV2, has been added to
lexer.h/.cpp, which provides a vector of structs, each struct
representing a single line of source file input, such that the struct
contains the line number and a vector of strings that are the tokens of
the line. This is more useful than being provided just a list of the
tokens themselves, because we need the line numbers to make compiler
error messages even remotely helpful.